Added much of the plugin functionality. On plugin activation, it'll move the object-cache.php file into the content directory. Once activated, it'll detect if the file is present. If not, it'll prompt to copy the plugin's copy to the directory. If one already exists, it'll prompt that it exists and give the option to overwrite it. On deactivation, the plugin will remove the file if no modifications have been made. This way, in case someone installs a new plugin that replaces the file, it won't remove that file.
I also added a few try/catch blocks in the cache-functions.php file. This way,
if the redis server is not running, or not present, it won't throw a fatal
error and crash the site. Instead, it'll report the error to the admin
dashboard.
