+++ composer.json
@@ -17,8 +17,27 @@
   "support": {
     "issues": "https://github.com/ericmann/wordpress-redis-backend/issues"
   },
+  "repositories": [
+    {
+      "type": "package",
+      "package": {
+        "name": "nathanielks/wp-admin-notice",
+        "version": "dev-master",
+        "type": "package",
+        "source": {
+          "type": "git",
+          "url": "https://gist.github.com/8910601.git",
+          "reference": "master"
+        },
+        "autoload": {
+          "psr-4": { "": "" }
+        }
+      }
+    }
+  ],
   "require": {
     "php": ">=5.3.2",
-    "predis/predis": ">=0.8.4"
+    "predis/predis": ">=0.8.4",
+    "nathanielks/wp-admin-notice": "dev-master"
   }
 }
